[CDAP-20648] Removed manager interfaces, removed resource name validation, addressed comments
